CSSANet/requirements.txt
Django==2.1.3

django-grappelli==2.12.1

django-filebrowser==3.11.1

Djangorestframework==3.9.0

django-rest-auth>=0.3.3

django-allauth

django_redis

psycopg2-binary

Pillow

gunicorn